% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{adequa_dados}
\alias{adequa_dados}
\title{Adequa dados observados e previstos}
\usage{
adequa_dados(
  resultados_leitura,
  v_usinas,
  fonte,
  modelo_nwp,
  data_inicio,
  data_fim
)
}
\arguments{
\item{resultados_leitura}{Lista de data.tables contendo os dados brutos,
incluindo observados, previstos e demais dados auxiliares.}

\item{v_usinas}{Vetor de identificadores das usinas a serem mantidas.}

\item{fonte}{Vetor de identificadores de fonte de observacao a serem filtrados.}

\item{modelo_nwp}{Vetor de identificadores de modelos NWP a serem filtrados.}

\item{data_inicio}{Data inicial no formato reconhecido por as.POSIXct.}

\item{data_fim}{Data final no formato reconhecido por as.POSIXct.}
}
\value{
Uma lista com as mesmas estruturas de entrada, mas filtradas de acordo com
os parametros de usinas, fontes, modelos e intervalo de tempo.
}
\description{
Aplica filtros sobre listas de dados observados e previstos de usinas,
retornando apenas os registros dentro das condicoes de interesse.
}
\details{
A funcao separa as listas de dados em tres grupos: observados, previstos
e auxiliares. Aplica filtros especificos a cada grupo e depois junta novamente
os resultados preservando os nomes originais.
}
\examples{
library(data.table)
horas <- seq.POSIXt(as.POSIXct("2025-05-26 00:00"), by = "1 hour", length.out = 4)
obs <- data.table(
    id_usina = c("U1", "U2"),
    id_fonte_observacao = "Consis",
    data_hora_observacao = rep(horas, each = 2),
    valor = runif(8)
)
prev <- data.table(
    id_modelo_nwp = "ModeloA",
    data_hora_previsao = horas,
    valor = runif(4)
)
resultados_leitura <- list(
    ger_obs = obs,
    dcorte_obs = obs,
    mhg = obs,
    mhg_sem_cortes = obs,
    irrad_prev = prev
)
dados_filtrados <- adequa_dados(
    resultados_leitura,
    v_usinas = "U1",
    fonte = "Consis",
    modelo_nwp = "ModeloA",
    data_inicio = "2025-05-26",
    data_fim = "2025-05-26 02:00:00"
)
lapply(dados_filtrados, head)

}
\seealso{
adequa_dados_observados, adequa_dados_previstos
}
