% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{adequa_dados_observados}
\alias{adequa_dados_observados}
\title{Filtra lista de dados observados}
\usage{
adequa_dados_observados(lista_obs, v_usinas, fonte, data_inicio, data_fim)
}
\arguments{
\item{lista_obs}{Lista de data.tables com dados observados.}

\item{v_usinas}{Vetor de identificadores de usinas.}

\item{fonte}{Vetor de identificadores de fonte de observacao.}

\item{data_inicio}{Data inicial no formato reconhecido por as.POSIXct.}

\item{data_fim}{Data final no formato reconhecido por as.POSIXct.}
}
\value{
Lista de data.tables filtrados.
}
\description{
Filtra lista de dados observados
}
