% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{adequa_dados_previstos}
\alias{adequa_dados_previstos}
\title{Filtra lista de dados previstos}
\usage{
adequa_dados_previstos(lista_prev, modelo_nwp, data_inicio, data_fim)
}
\arguments{
\item{lista_prev}{Lista de data.tables com dados previstos.}

\item{modelo_nwp}{Vetor de identificadores de modelos NWP.}

\item{data_inicio}{Data inicial no formato reconhecido por as.POSIXct.}

\item{data_fim}{Data final no formato reconhecido por as.POSIXct.}
}
\value{
Lista de data.tables filtrados.
}
\description{
Filtra lista de dados previstos
}
