% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{associa_nwp_usina}
\alias{associa_nwp_usina}
\title{Associa Coordenadas de Previsao NWP a Cada Usina}
\usage{
associa_nwp_usina(dt_usinas, dt_irrad_prev)
}
\arguments{
\item{dt_usinas}{data.table contendo pelo menos as colunas:
\code{id_usina}, \code{latitude} e \code{longitude}.}

\item{dt_irrad_prev}{data.table contendo previsoes de irradiancia,
com colunas \code{latitude} e \code{longitude} (alem de outras colunas de dados).}
}
\value{
data.table com todos os registros de \code{dt_irrad_prev} filtrados
para a coordenada mais proxima de cada usina e com a coluna
\code{id_usina} adicionada.
}
\description{
Para cada usina em \code{dt_usinas}, encontra a coordenada mais proxima
na previsao NWP (\code{dt_irrad_prev}) usando distancia Haversine e filtra
os dados correspondentes. Retorna os dados de irradiancia previstos
com a coluna \code{id_usina} associada.
}
\details{
A funcao:
\enumerate{
\item Calcula a distancia Haversine entre cada usina e todas as coordenadas
unicas da previsao (com memoizacao por coordenadas).
\item Seleciona a coordenada mais proxima para cada usina.
\item Filtra os dados da previsao para essa coordenada.
\item Adiciona \code{id_usina} e reorganiza as colunas.
}
}
