% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{clear_haversine_cache}
\alias{clear_haversine_cache}
\title{Limpa Cache de Distancias Haversine}
\usage{
clear_haversine_cache()
}
\value{
\code{invisible(NULL)}
}
\description{
Remove todos os mapeamentos de coordenadas NWP-usina armazenados em cache.
Use quando as coordenadas das usinas ou do grid NWP mudarem.
}
