% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-linear-regression.r
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Ajusta Modelo de Regressao Linear}
\usage{
fit_linear_regression(dty, dtx, dty_bruta, ...)
}
\arguments{
\item{dty}{data.table com dados de geracao observada (variavel resposta)}

\item{dtx}{data.table com dados de irradiacao prevista (variavel explicativa)}

\item{dty_bruta}{data.table com dados de geracao bruta (sem cortes)}

\item{...}{argumentos adicionais (ignorados)}
}
\value{
objeto S3 com classe \code{"linear_regression_model"} contendo:
\describe{
\item{\code{parametros}}{\code{data.frame} com coeficientes \code{a} (angular) e \code{b}
(sempre zero) indexados por horario \code{"HH:MM"}}
}
}
\description{
Ajusta regressoes lineares sem intercepto por horario de meia hora entre
geracao observada e irradiacao prevista. Retorna objeto com classe
\code{"linear_regression_model"}.
}
\seealso{
\code{\link[=fit_model]{fit_model()}}, \code{\link[=predict_model.linear_regression_model]{predict_model.linear_regression_model()}},
\code{\link[=model_metadata.linear_regression_model]{model_metadata.linear_regression_model()}}
}
