% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_corte_observado}
\alias{get_corte_observado}
\title{Leitura De Corte Observado}
\usage{
get_corte_observado(usina = NULL, fonte = NULL, input_dir = NULL)
}
\arguments{
\item{usina}{vetor de codigos das usina que devem ser buscadas}

\item{fonte}{vetor de fontes de dados a serem buscados}
}
\value{
\code{data.table} de corte observado padronizado nas colunas:
* \code{id_fonte_observacao}: string indicando a fonte de dados
* \code{id_usina}: string indicando codigo da usina
* \code{data_hora_observacao}: POSIX em UTC indicando horario da observacao
* \code{valor}: valor da corte observado
* \code{status}: vazio, nao existe para este dado
}
\description{
Wrapper para leitura de corte observado e subsets por usina ou fonte de dados
}
