% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_dados_historicos}
\alias{get_dados_historicos}
\title{Leitura dos dados}
\usage{
get_dados_historicos(v_usinas, fonte, input_dir, modelo_nwp)
}
\arguments{
\item{fonte}{vetor de fontes de dados a serem buscados}

\item{usina}{vetor de codigos das usina que devem ser buscadas}
}
\value{
\code{data.table} de geracao observada padronizado nas colunas:
* \code{id_fonte_observacao}: string indicando a fonte de dados
* \code{id_usina}: string indicando codigo da usina
* \code{data_hora_observacao}: POSIX em UTC indicando horario da observacao
* \code{valor}: valor da geracao observada
* \code{status}: vazio, nao existe para este dado
}
\description{
Wrapper para leitura dos dados
}
