% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_irradiancia_prevista}
\alias{get_irradiancia_prevista}
\title{Leitura De Irradiancia Prevista}
\usage{
get_irradiancia_prevista(modelo_nwp = NULL, input_dir = NULL)
}
\arguments{
\item{modelo_nwp}{vetor de modelos a serem buscados}
}
\value{
\code{data.table} de irradiancia prevista padronizado nas colunas:
* \code{id_modelo_nwp}: string indicando o modelo de previsao
* \code{latitude}: latutide da usina
* \code{longitude}: longitude da usina
* \code{data_hora_rodada}: POSIX em UTC indicando horario da rodada do modelo
* \code{data_hora_previsao}: POSIX em UTC indicando horario da previsao
* \code{valor}: valor da irradiancia prevista
}
\description{
Wrapper para leitura de irradiancia prevista e subsets por modelo
}
