% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_melhor_historico_geracao_sem_cortes}
\alias{get_melhor_historico_geracao_sem_cortes}
\title{Leitura De Melhor Historico de Geracao sem Cortes}
\usage{
get_melhor_historico_geracao_sem_cortes(
  usina = NULL,
  fonte = NULL,
  input_dir = NULL
)
}
\arguments{
\item{usina}{vetor de codigos das usina que devem ser buscadas}

\item{fonte}{vetor de fontes de dados a serem buscados}
}
\value{
\code{data.table} de potencia disponivel observada padronizado nas colunas:
* \code{id_fonte_observacao}: string indicando a fonte de dados
* \code{id_usina}: string indicando codigo da usina
* \code{data_hora_observacao}: POSIX em UTC indicando horario da observacao
* \code{valor}: valor da potencia disponivel observada
* \code{status}: vazio, nao existe para este dado
}
\description{
Wrapper para leitura de melhor historico de geracao sem cortes e subsets por usina ou modelo
}
