% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.r
\name{get_parser}
\alias{get_parser}
\title{Constroi o interpretador de argumentos de linha de comando}
\usage{
get_parser()
}
\value{
\code{ArgumentParser} com os argumentos suportados pelo modelo
}
\description{
Constroi o interpretador de argumentos de linha de comando
}
