% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_usinas}
\alias{get_usinas}
\title{Leitura De Dados Das Usinas}
\usage{
get_usinas(usina = NULL, input_dir = NULL)
}
\arguments{
\item{usina}{vetor de codigos das usina que devem ser buscadas}
}
\value{
\code{data.table} de dados das usinas padronizado nas colunas:
* \code{id_usina}: string indicando codigo da usina
* \code{latitude}: latutide da usina
* \code{longitude}: longitude da usina
* \code{capacidade_instalada_MW}: capacidade instalada da usina
* \code{data_inicio_operacao_comercial}: POSIX UTC indicando inicio entrada em operacao
}
\description{
Wrapper para leitura de dados das usinas e subset por usinas
}
