% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.r
\name{inner_parser_generic_args}
\alias{inner_parser_generic_args}
\title{Auxiliar para adicionar argumentos genericos}
\usage{
inner_parser_generic_args(parser)
}
\arguments{
\item{parser}{\code{ArgumentParser} a ser modificado

Funcao interna, nao deve ser chamada diretamente pelo usuario}
}
\description{
Auxiliar para adicionar argumentos genericos
}
