% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{inner_reader}
\alias{inner_reader}
\title{Auxiliar Para Leitura De Dados}
\usage{
inner_reader(usina = NULL, fonte = NULL, table = "", input_dir = ".")
}
\description{
Funcao interna, nao deve ser chamada diretamente pelo usuario
}
