% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{inner_reader_nwp}
\alias{inner_reader_nwp}
\title{Auxiliar Para Leitura de Dados Previstos}
\usage{
inner_reader_nwp(modelo_nwp = NULL, table = "", input_dir = ".")
}
\arguments{
\item{table}{nome-base da tabela (sem extensao .csv)}

\item{dt}{\code{data.table} lido}

\item{output_dir}{diretorio onde o arquivo sera salvo}
}
\value{
data.table lido
}
\description{
Funcao interna para ler data.tables em CSV. Nao deve ser chamada diretamente pelo usuario.
}
