% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia-dados.r
\name{marca_zeros_consecutivos}
\alias{marca_zeros_consecutivos}
\title{Marca zeros consecutivos como NA}
\usage{
marca_zeros_consecutivos(v, n_lim)
}
\arguments{
\item{v}{Vetor numerico a ser avaliado.}

\item{n_lim}{Numero minimo de zeros consecutivos necessario para que a
sequencia seja substituida por NA.}
}
\value{
O vetor v com NA nas posicoes onde ocorreram sequencias de zeros
com comprimento maior ou igual ao limite especificado.
}
\description{
Esta funcao identifica sequencias de zeros em um vetor numerico e substitui
por NA todas as posicoes pertencentes a sequencias cujo comprimento seja
maior ou igual ao limite definido.
}
\details{
A funcao utiliza rle para identificar sequencias consecutivas de zeros.
Em seguida, constrói um vetor logico por meio de inverse.rle indicando quais
posicoes pertencem a sequencias longas o suficiente para serem marcadas.
Finalmente, as posicoes selecionadas sao substituidas por NA_real_.
}
