% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini-MH.r
\docType{package}
\name{melhorhistoricosolar-package}
\alias{melhorhistoricosolar}
\alias{melhorhistoricosolar-package}
\title{\code{melhor_historico_solar}}
\description{
Construcao Do Melhor Historico De Insumos Observados Para Previsao De Solar Fotovoltaica
}
\details{
Este pacote contem as funcionalidades necessarias para tratamento de dados de geracao,
cortes, irradiancia e potencia disponivel observados e posterior construcao dos melhores
historicos de geracao e geracao sem cortes
}
\author{
\strong{Maintainer}: Paulo Nascimento \email{pnascimento@ons.org.br}

}
\keyword{internal}
