% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict_main}
\alias{predict_main}
\title{Previsao De Melhor Historico Solar}
\usage{
predict_main(args)
}
\arguments{
\item{args}{lista de argumentos para execucao do melhor historico. Veja Detalhes}
}
\value{
escreve no diretorio determinado em \code{args} os historicos produzidos
}
\description{
Funcao principal para realizacao dos melhores historicos de solar fotovoltaica
}
