% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.r
\name{read_env_flag}
\alias{read_env_flag}
\title{Le variavel de ambiente booleana}
\usage{
read_env_flag(name, default = FALSE)
}
\arguments{
\item{name}{nome da variavel de ambiente}

\item{default}{valor logico retornado quando a variavel nao esta definida
ou possui valor invalido. Padrao: \code{FALSE}}
}
\value{
\code{TRUE} ou \code{FALSE}
}
\description{
Interpreta o valor de uma variavel de ambiente como logico. Aceita
\code{"true"}, \code{"1"}, \code{"yes"} como \code{TRUE} e \code{"false"}, \code{"0"}, \code{"no"} como
\code{FALSE}. Valores invalidos emitem aviso e retornam o padrao.
}
