% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{read_model_artifact}
\alias{read_model_artifact}
\title{Leitura do artefato de modelo treinado para uso em previsao}
\usage{
read_model_artifact(iu, artifact_dir = ".")
}
\arguments{
\item{iu}{ID da usina, usado para montar o nome do arquivo}

\item{artifact_dir}{diretorio de saida onde foi salvo o artefato}
}
\value{
Objeto em R representando o modelo treinado
}
\description{
Le o conteudo do objeto do modelo de um arquivo RDS
}
