% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{reset_parallel_plan}
\alias{reset_parallel_plan}
\title{Restaura Plano de Execucao Paralela Anterior}
\usage{
reset_parallel_plan(old_plan)
}
\arguments{
\item{old_plan}{plano anterior retornado por \code{\link[=setup_parallel_plan]{setup_parallel_plan()}}}
}
\value{
\code{invisible(NULL)}
}
\description{
Restaura um plano de execucao paralela salvo previamente, tipicamente
obtido como retorno de \code{\link[=setup_parallel_plan]{setup_parallel_plan()}}.
}
\examples{
old <- setup_parallel_plan(strategy = "sequential")
reset_parallel_plan(old)

}
