% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{split_args_by_plant}
\alias{split_args_by_plant}
\title{Divide Argumentos por Usina para Despacho Paralelo}
\usage{
split_args_by_plant(extra_args, v_usinas)
}
\arguments{
\item{extra_args}{lista nomeada de argumentos, como construida por
\code{predict_main} ou \code{train_main}}

\item{v_usinas}{vetor de IDs de usinas a processar}
}
\value{
lista nomeada por \code{id_usina}, onde cada elemento e uma lista
com a mesma estrutura de \code{extra_args}, porem com data.tables filtrados
para a usina correspondente
}
\description{
Particiona os data.tables contidos em \code{extra_args} por \code{id_usina},
gerando uma lista de argumentos por usina pronta para ser enviada a
workers paralelos. Elementos que nao sao data.tables ou que nao possuem
a coluna \code{id_usina} sao replicados inalterados para cada usina.
}
