% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valid_tipos_unit}
\alias{valid_tipos_unit}
\title{Auxiliar De \code{valid_tipos}}
\usage{
valid_tipos_unit(x, tipos)
}
\arguments{
\item{x}{escalar, elemento de uma chave do arquivo de configuracao}

\item{tipos}{tipos esperados de \code{x}, escalar ou lista}
}
\value{
booleano indicando se validacao encerrou com sucesso ou nao
}
\description{
Funcao interna para isolar o loop ao longo de \code{l} em \code{valid_tipos}
}
