% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_corte_observado}
\alias{valida_corte_observado}
\title{Funcao para validar o \code{data.table} de corte observado}
\usage{
valida_corte_observado(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de corte observado
}
