% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_geracao_observada}
\alias{valida_geracao_observada}
\title{Funcao para validar o \code{data.table} de geracao observada}
\usage{
valida_geracao_observada(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de geracao observada
}
