% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_intervalo}
\alias{valida_intervalo}
\title{Auxiliar para validar intervalo numerico}
\usage{
valida_intervalo(dados, nome_coluna, limites)
}
\arguments{
\item{dados}{\code{data.table} de dados para validacao}

\item{nome_coluna}{nome da coluna a ser validada}

\item{limites}{lista com limites inferior e superior dos dados da coluna}
}
\value{
lista com nomes 'sucesso' e 'mensagem_erro' para tratamento externo
}
\description{
Auxiliar para validar intervalo numerico
}
