% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_irradiancia_prevista}
\alias{valida_irradiancia_prevista}
\title{Funcao para validar o \code{data.table} de irradiancia prevista}
\usage{
valida_irradiancia_prevista(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de irradiancia prevista
}
