% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_limites}
\alias{valida_limites}
\title{Auxiliar para validacoes genericas de colunas numericas}
\usage{
valida_limites(dados, lista_limites)
}
\arguments{
\item{dados}{\code{data.table} de dados para validacao}

\item{lista_limites}{lista nomeada com os limites de cada coluna}
}
\description{
Auxiliar para validacoes genericas de colunas numericas
}
