% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_melhor_historico_geracao}
\alias{valida_melhor_historico_geracao}
\title{Funcao para validar o \code{data.table} de melhor historico de geracao}
\usage{
valida_melhor_historico_geracao(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de melhor historico de geracao
}
