% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_nomes_colunas}
\alias{valida_nomes_colunas}
\title{Auxiliar para validar um \code{data.table} em termos dos nomes das colunas}
\usage{
valida_nomes_colunas(dados, lista_nomes)
}
\arguments{
\item{dados}{\code{data.table} de dados para validacao}

\item{lista_nomes}{vetor de nomes de colunas para serem validados}
}
\value{
lista com nomes 'sucesso' e 'mensagem_erro' para tratamento externo
}
\description{
Auxiliar para validar um \code{data.table} em termos dos nomes das colunas
}
