% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valida_nomes_config}
\alias{valida_nomes_config}
\title{Valida Chaves Do Arquivo De Configuracao}
\usage{
valida_nomes_config(config)
}
\arguments{
\item{config}{lista de configuracoes}
}
\value{
NULL se config possui todas as chaves; levanta erro do contrario
}
\description{
Checa se um arquivo de configuracao lido possui todas as chaves necessarias
}
