% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_nomes_tipos}
\alias{valida_nomes_tipos}
\title{Auxiliar para validacoes genericas de nomes e tipos}
\usage{
valida_nomes_tipos(dados, lista_nomes, lista_tipos)
}
\arguments{
\item{dados}{\code{data.table} de dados para validacao}

\item{lista_nomes}{vetor de nomes de colunas para serem validados}

\item{lista_tipos}{lista nomeada com os tipos de dados esperados em cada coluna}
}
\description{
Auxiliar para validacoes genericas de nomes e tipos
}
