% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_potencia_disponivel_observada}
\alias{valida_potencia_disponivel_observada}
\title{Funcao para validar o \code{data.table} de potencia disponivel observada}
\usage{
valida_potencia_disponivel_observada(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de potencia disponivel observada
}
