% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valida_tipos_config}
\alias{valida_tipos_config}
\title{Valida Tipos Das Chaves Do Arquivo De Configuracao}
\usage{
valida_tipos_config(config)
}
\arguments{
\item{config}{lista de configuracoes}
}
\value{
NULL se config possui todos os tipos corretos; levanta erro do contrario
}
\description{
Checa se valores das chaves no arquivo lido sao dos tipos corretos
}
