% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_tipos_dados_colunas}
\alias{valida_tipos_dados_colunas}
\title{Auxiliar para validar um \code{data.table} em termos dos tipos de dados das colunas}
\usage{
valida_tipos_dados_colunas(dados, lista_tipos)
}
\arguments{
\item{dados}{\code{data.table} de dados para validacao}

\item{lista_tipos}{lista nomeada com os tipos de dados esperados em cada coluna}
}
\value{
lista com nomes 'sucesso' e 'mensagem_erro' para tratamento externo
}
\description{
Auxiliar para validar um \code{data.table} em termos dos tipos de dados das colunas
}
