% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_usinas}
\alias{valida_usinas}
\title{Funcao para validar o \code{data.table} de usinas}
\usage{
valida_usinas(dados)
}
\arguments{
\item{dados}{\code{data.table} para validacao}
}
\description{
Funcao para validar o \code{data.table} de usinas
}
