% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{write_model_artifact}
\alias{write_model_artifact}
\title{Escrita do artefato de modelo treinado para uso em previsao}
\usage{
write_model_artifact(model, iu, artifact_dir = ".")
}
\arguments{
\item{model}{objeto em R representando o modelo treinado}

\item{iu}{ID da usina, usado para nomear o arquivo}

\item{artifact_dir}{diretorio de saida onde sera salvo o artefato}
}
\value{
Caminho completo do arquivo salvo
}
\description{
Salva o conteudo do objeto do modelo em um arquivo RDS
}
