% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia-dados.r
\name{checa_valores_overbound}
\alias{checa_valores_overbound}
\title{Remove Valores Fora de Limites Pre-Estabelecidos}
\usage{
checa_valores_overbound(dt, limites = c(0, Inf))
}
\arguments{
\item{dt}{Um data.table contendo ao menos uma coluna chamada valor, com os dados numericos a serem verificados.}

\item{limites}{Vetor numerico de comprimento 2, indicando o limite inferior e superior permitidos, respectivamente.
Valores fora desse intervalo serao considerados invalidos e substituidos por NA.}
}
\value{
O mesmo data.table de entrada, com os valores da coluna valor fora dos limites substituidos por NA.
}
\description{
Verifica e substitui por NA os valores que estao fora de um intervalo definido de limites inferior e superior.
}
\details{
A funcao aplica uma verificacao simples nos dados numericos da coluna valor.
Todo valor menor que o limite inferior ou maior que o limite superior definido no argumento limites e substituido por NA.
}
\seealso{
remove_congelados, checa_valores_congelados
}
