% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{combina_dados_tempo}
\alias{combina_dados_tempo}
\title{Combina Dados em Serie Temporal}
\usage{
combina_dados_tempo(dt1, dt2)
}
\arguments{
\item{dt1}{data.table contendo pelo menos as colunas \code{id_usina},
\code{data_hora_observacao}, \code{valor} e \code{status}.
Representa a base de dados principal.}

\item{dt2}{data.table contendo as mesmas colunas que \code{dt1}.
Representa os dados que devem sobrepor a base.}
}
\value{
Um data.table com as colunas:
- \code{id_fonte_observacao}: sempre preenchido com "Consis"
- \code{id_usina}
- \code{data_hora_observacao}
- \code{valor}
- \code{status}
Os dados estao ordenados por \code{id_usina} e \code{data_hora_observacao}.
}
\description{
Combina dois data.tables , alinhando a serie temporal
em intervalos de 30 minutos e aplicando sobreposicao de valores.
Caso existam valores em \code{dt2}, eles sobrescrevem os de \code{dt1}.
}
\details{
A funcao cria uma grade temporal completa entre a menor e a maior
data das duas entradas, com intervalo de 30 minutos, para todos os
usinas presentes em \code{dt1} e \code{dt2}. Em seguida:
\enumerate{
\item Faz merge com \code{dt1} como base.
\item Faz merge com \code{dt2} e aplica sobreposicao dos valores e status.
\item Adiciona a coluna \code{id_fonte_observacao} com valor "Consis".
}
}
