% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia-dados.r
\name{consiste_geracao_unit}
\alias{consiste_geracao_unit}
\title{Trata E Combina Dados De Geracao}
\usage{
consiste_geracao_unit(
  dados_usina,
  geracao_usina,
  ordem_prioridade,
  limite_dados
)
}
\arguments{
\item{dados_usina}{\code{data.table} de dados das usinas com apenas a linha da usina a ser tratada}

\item{geracao_usina}{\code{data.table} de geracao observada apenas da usina a ser tratada}
}
\value{
\code{geracao_usina} consolidado em apenas uma fonte denominada \code{"Consis"}
}
\description{
Condensa um dado de geracao observada com multiplas fontes em uma unica, combinada
}
