% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{create_metrics}
\alias{create_metrics}
\title{Cria Registro de Metricas de Execucao}
\usage{
create_metrics(run_id, mode)
}
\arguments{
\item{run_id}{character escalar, identificador da execucao}

\item{mode}{character escalar, \code{"train"} ou \code{"predict"}}
}
\value{
lista com estrutura de metricas vazia
}
\description{
Cria Registro de Metricas de Execucao
}
