% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{get_pkg_logger}
\alias{get_pkg_logger}
\title{Getter Do Objeto Logger}
\usage{
get_pkg_logger()
}
\value{
objeto \code{Logger} como retornado por \code{lgr::get_logger()}
}
\description{
Getter Do Objeto Logger
}
