% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.r
\name{inner_parser_specific_args}
\alias{inner_parser_specific_args}
\title{Auxiliar para adicionar argumentos especificos}
\usage{
inner_parser_specific_args(parser)
}
\description{
Funcao interna, nao deve ser chamada diretamente pelo usuario
}
