% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{inner_writer}
\alias{inner_writer}
\title{Auxiliar Para Escrita de Dados}
\usage{
inner_writer(dt, table = "", output_dir = ".", output_type = "csv")
}
\arguments{
\item{dt}{\code{data.table} a ser salvo}

\item{table}{nome-base da tabela (sem extensao .csv)}

\item{output_dir}{diretorio onde o arquivo sera salvo}

\item{output_type}{tipo de arquivo a ser salvo ("csv" ou "parquet")}
}
\value{
Caminho do arquivo salvo
}
\description{
Funcao interna para salvar data.tables em CSV. Nao deve ser chamada diretamente pelo usuario.
}
