% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-linear-regression.r
\name{linear_regression_strategy}
\alias{linear_regression_strategy}
\title{Construtor de Estrategia de Regressao Linear}
\usage{
linear_regression_strategy(...)
}
\arguments{
\item{...}{parametros adicionais repassados a \code{\link[=new_model_strategy]{new_model_strategy()}}}
}
\value{
objeto S3 com classe \code{c("linear_regression", "model_strategy")}
}
\description{
Cria um objeto \code{model_strategy} especializado para regressao linear,
com classe dupla \code{c("linear_regression", "model_strategy")}.
Os metodos \code{\link[=fit_model]{fit_model()}}, \code{\link[=predict_model]{predict_model()}} e \code{\link[=model_metadata]{model_metadata()}} delegam
para as funcoes existentes \code{\link[=ajusta_regressao_ger_irrad]{ajusta_regressao_ger_irrad()}} e
\code{\link[=substitui_por_estimativas]{substitui_por_estimativas()}}.
}
\examples{
s <- linear_regression_strategy()
inherits(s, "linear_regression")
inherits(s, "model_strategy")

}
\seealso{
\code{\link[=new_model_strategy]{new_model_strategy()}}, \code{\link[=fit_model]{fit_model()}}, \code{\link[=predict_model]{predict_model()}},
\code{\link[=model_metadata]{model_metadata()}}
}
