% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{organiza_resultados}
\alias{organiza_resultados}
\title{Organiza Resultados de Previsao por Usina}
\usage{
organiza_resultados(resultados, v_usinas)
}
\arguments{
\item{resultados}{Lista contendo, para cada usina, um sub-lista com dois elementos:
\itemize{
\item \code{com_cortes}: data.table com os dados considerando os efeitos de corte.
\item \code{sem_cortes}: data.table com os dados sem considerar os cortes.
}}

\item{v_usinas}{Vetor de caracteres com os IDs das usinas, na mesma ordem da lista \code{resultados}.}
}
\value{
Uma lista com dois data.tables:
\itemize{
\item \code{com_cortes}: dados de todas as usinas, concatenados e com a coluna \code{id_usina} preenchida.
\item \code{sem_cortes}: dados das mesmas usinas sem considerar cortes, tambem com \code{id_usina}.
}
}
\description{
Agrupa os resultados processados individualmente por usina em dois data.tables:
um com consideracao de cortes e outro sem.
}
\details{
A funcao percorre os elementos da lista \code{resultados}, adiciona a identificacao da usina correspondente,
e empacota os dados finais em dois data.tables: um com cortes e outro sem. Util para consolidar os resultados
apos o processamento individual de cada usina.
}
\seealso{
processar_usina, predict_main
}
