% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parse_config}
\alias{parse_config}
\title{Interpreta Arquivo De Configuracao}
\usage{
parse_config(config, conn)
}
\arguments{
\item{config}{lista nomeada de}

\item{conn}{objeto de conexao com um banco}
}
\value{
lista de argumentos interpretados
}
\description{
Interpreta Arquivo De Configuracao
}
