% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_ids_usinas}
\alias{parsearg_ids_usinas}
\title{Interpretador De Chave \code{ids_usinas}}
\usage{
parsearg_ids_usinas(x, conn)
}
\arguments{
\item{x}{valor da chave \code{ids_usinas}; lista vazia ou de codigos de usinas}

\item{conn}{objeto de conexao com um banco}
}
\value{
se \code{x} era uma lista vazia, retorna um vetor com todos os ids no banco \code{conn}; do
contrario retorna \code{x} vetorizado
}
\description{
Funcao interna de \code{\link{parse_config}} para interpretar o parametro \code{ids_usinas} da configuracao
}
