% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_janela}
\alias{parsearg_janela}
\alias{parsearg_janela.numeric}
\alias{parsearg_janela.character}
\title{Interpretador De Chave \code{janela}}
\usage{
parsearg_janela(x)

\method{parsearg_janela}{numeric}(x)

\method{parsearg_janela}{character}(x)
}
\arguments{
\item{x}{valor da chave \code{janela}; numerico ou vetor de duas strings de data}
}
\value{
vetor \code{Date} de duas posicoes indicando inicio e fim da janela de melhor historico
}
\description{
Funcao interna de \code{\link{parse_config}} para interpretar o parametro \code{janela} da configuracao
}
