% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia-dados.r
\name{remove_congelados}
\alias{remove_congelados}
\title{Remove Valores Congelados de um Vetor Numerico}
\usage{
remove_congelados(v, n_valores, limiar)
}
\arguments{
\item{v}{Vetor numerico com os dados de geracao observada.}

\item{n_valores}{Numero de valores consecutivos considerados para a janela de deteccao.
Define o tamanho da sequencia minima a ser avaliada como "congelada".}

\item{limiar}{Variacao maxima permitida entre os valores dentro da janela.
Se a variacao dentro da janela for menor ou igual a esse valor, ela sera considerada congelada.}
}
\value{
Vetor numerico com os mesmos valores de entrada, mas com os trechos identificados como congelados substituidos por NA.
}
\description{
Identifica e substitui por NA os trechos de dados considerados "congelados",
ou seja, sequencias de valores com variacao insignificante dentro de uma janela deslizante.
}
\details{
A funcao percorre o vetor original com uma janela deslizante de tamanho n_valores.
Para cada janela, calcula-se a diferenca entre o valor maximo e o minimo.
Se essa diferenca for menor ou igual ao limiar, os valores dentro da janela sao marcados como congelados e substituidos por NA.
}
\examples{
v <- c(10, 10, 10, 10, 10, 15, 20, 25)
remove_congelados(v, n_valores = 5, limiar = 0.01)

}
\seealso{
checa_valores_congelados
}
