% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valid_tipos}
\alias{valid_tipos}
\title{Validacao Singular De Uma Chave}
\usage{
valid_tipos(l, tipos)
}
\arguments{
\item{l}{valor de uma chave do arquivo de configuracao, escalar ou lista}

\item{tipos}{tipos esperados de \code{l}, escalar ou lista}
}
\value{
booleano indicando se validacao encerrou com sucesso ou nao
}
\description{
Funcao interna auxiliar de \code{valida_tipos_config()}
}
\details{
Tanto \code{l} quanto \code{tipos} podem ser escalares ou listas. No caso de \code{l}, cada elemento sera checado
individualmente. Se \code{tipos} for uma lista, \code{l} sera checado contra cada um dos tipos e retorna
\code{TRUE} se ao menos um deles for valido
}
