% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{write_melhor_historico_geracao}
\alias{write_melhor_historico_geracao}
\title{Escrita de Melhor Historico de Geracao}
\usage{
write_melhor_historico_geracao(dt, output_dir = ".")
}
\arguments{
\item{dt}{\code{data.table} com os dados a serem salvos}

\item{output_dir}{diretorio de saida onde sera salvo o arquivo Parquet}
}
\value{
vazio, apenas escreve arquivo
}
\description{
Salva os dados de melhor historico de geracao em disco, formatados e validados
}
