% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{write_melhor_historico_geracao_sem_cortes}
\alias{write_melhor_historico_geracao_sem_cortes}
\title{Escrita de Melhor Historico de Geracao sem Cortes}
\usage{
write_melhor_historico_geracao_sem_cortes(dt, output_dir = ".")
}
\arguments{
\item{dt}{\code{data.table} com os dados a serem salvos}

\item{output_dir}{diretorio de saída onde sera salvo o arquivo CSV}
}
\value{
Caminho completo do arquivo salvo
}
\description{
Salva os dados de melhor historico de geracao sem cortes em disco, formatados e validados
}
